Refactor upload components and update imports for ArtistArtUpload in ArtistForm
