+++ app/upload/page.tsx
@@ -1,10 +1,10 @@
 "use client";
-import { TrackUpload } from "@/components/track-upload";
+import { TrackUpload } from "@/components/Upload";
 import * as React from "react";
 
-export interface IUploadTrackPageProps {}
+export interface IUploadTrackPageProps { }
 
-export default function UploadTrackPage({}: IUploadTrackPageProps) {
+export default function UploadTrackPage({ }: IUploadTrackPageProps) {
   return (
     <div>
       <TrackUpload />
