+++ components/Upload/ArtistArtUpload/ArtistAvatarUpload.tsx
