+++ components/Upload/ArtistArtUpload/ArtistBannerUpload.tsx
