+++ components/Upload/TrackUpload/MetadataDisplay.tsx
